+++ x/epochs/types/identifier.go
@@ -1,6 +1,7 @@
 package types
 
 import (
+	"errors"
 	"fmt"
 )
 
@@ -19,6 +20,23 @@ func ValidateEpochIdentifierInterface(i interface{}) error {
 func ValidateEpochIdentifierString(s string) error {
 	if s == "" {
 		return fmt.Errorf("empty distribution epoch identifier: %+v", s)
+	}
+	return nil
+}
+
+// Validate also validates epoch info.
+func (epoch EpochInfo) Validate() error {
+	if epoch.Identifier == "" {
+		return errors.New("epoch identifier should NOT be empty")
+	}
+	if epoch.Duration == 0 {
+		return errors.New("epoch duration should NOT be 0")
+	}
+	if epoch.CurrentEpoch < 0 {
+		return errors.New("epoch CurrentEpoch must be non-negative")
+	}
+	if epoch.CurrentEpochStartHeight < 0 {
+		return errors.New("epoch CurrentEpoch must be non-negative")
 	}
 	return nil
 }